# Optimize models.MakeKey