Optimizes `models.MakeKey` this was showing up in some profiles for writes.

```
benchcmp /tmp/old.txt /tmp/new.txt
benchmark              old ns/op     new ns/op     delta
BenchmarkMakeKey-8     1139          177           -84.46%

benchmark              old allocs     new allocs     delta
BenchmarkMakeKey-8     7              2              -71.43%

benchmark              old bytes     new bytes     delta
BenchmarkMakeKey-8     272           96            -64.71%
```